pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "robotframework-mbt"
version = "0.11.0"
description = "Model-Based Testing in Robot framework with test case generation"
readme = "README.md"
authors = [{ name = "Johan Foederer", email = "github@famfoe.nl" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
keywords = ["robotframework", "robot", "mbt", "bdd", "testing"]
dependencies = [
    "robotframework >= 7.3",
]
requires-python = ">=3.10"

[tool.setuptools.packages.find]
include = ["robotmbt*"]

[project.urls]
Homepage = "https://github.com/JFoederer/robotframeworkMBT"

[project.optional-dependencies]
visualisation = ["networkx", "bokeh", "grandalf", "jsonpickle"]
